relojito/relojito/requirements/development.txt
coverage==3.7.1
django-debug-toolbar==1.0.1
django-extensions==1.3.3
enum34==0.9.22
factory-boy==2.3.1
ipaddress==1.0.6
ipdb==0.8
ipython==1.2.1
isort==3.6.2
natsort==3.1.0
pies2overrides==2.6.0
pies==2.6.0
pip-tools==0.3.4
sqlparse==0.1.10
Werkzeug==0.9.4
